% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart.R
\name{bar_chart_spec}
\alias{bar_chart_spec}
\title{Title}
\usage{
bar_chart_spec(
  gr,
  chromosome_field = "seqnames",
  genomic_fields = c("start", "end"),
  width = 800,
  height = 180,
  y_field = "peak",
  bin_size = 5,
  title = "Basic Marks: bar",
  subtitle = "Tutorial Examples"
)
}
\arguments{
\item{gr}{}

\item{chromosome_field}{}

\item{genomic_fields}{}

\item{width}{}

\item{height}{}

\item{y_field}{}

\item{bin_size}{}

\item{title}{}

\item{subtitle}{}
}
\description{
Title
}
\examples{

# generate example GRanges object
library(GenomicRanges)
set.seed(42)

# number of rows per chromosome
n <- 1000  

gr <- GRanges(
  seqnames = rep(paste0("chr", c(1:22, 'X', 'Y')), each = n),
  ranges = IRanges(
    start = rep(sample(1:250000000, n, replace = TRUE), each = 24),
    width = round(600 + rexp(n, rate = 1/1800)),
    peak = round(50 + rexp(n, rate = 1/500))
  )
)

# generate gosling spec and render visualization
spec <- bar_chart_spec(gr)
print(spec)

goslingr(spec)
}
