% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{create_gr_data}
\alias{create_gr_data}
\title{Create gosling data spec from GRanges object}
\usage{
create_gr_data(gr)
}
\arguments{
\item{gr}{GRanges object}
}
\description{
Create gosling data spec from GRanges object
}
