% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goslingr.R
\name{goslingr}
\alias{goslingr}
\title{Title TODO}
\usage{
goslingr(spec, width = NULL, height = NULL)
}
\arguments{
\item{spec}{Gosling specification either as an R list or as a json string.}
}
\description{
Description TODO
}
