% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_gosling}
\alias{install_gosling}
\title{Install Gos Python package}
\usage{
install_gosling(
  envname = "r-gosling",
  version = NULL,
  new_env = identical(envname, "r-gosling"),
  higlass_server = TRUE,
  ...
)
}
\arguments{
\item{envname}{\code{character}, name of python environment into which to install (default is \code{'r-gosling'})}

\item{version}{\code{character}, version of Gos to install. For general use of this package,}

\item{new_env}{\code{boolean}, should the existing python environment recreated before installation.}

\item{higlass_server}{\code{boolean}, if \code{TRUE} (default), installs \code{clodius} and \code{servir} so that
local files will be served by a background Higlass server.}

\item{...}{other arguments sent to \code{reticulate::py_install()}}
}
\value{
invisible \code{NULL}, called for side-effects
}
\description{
This function wraps installation functions from \link[reticulate:reticulate]{reticulate}
to install the Python package \strong{gos}.
}
\details{
This package uses the \link[reticulate:reticulate]{reticulate} package
to make an interface with the \href{https://github.com/gosling-lang/gos}{Gos}
Python package. To promote consistency in usage of \strong{reticulate} among
different R packages, it is
\href{https://rstudio.github.io/reticulate/articles/package.html#installing-python-dependencies}{recommended}
to use a common Python environment, called \code{"r-reticulate"}.

Depending on your setup, you can create this environment using
\code{reticulate::conda_create()} or \code{\link[reticulate:virtualenv-tools]{reticulate::virtualenv_create()}},
as described in this
\href{https://rstudio.github.io/reticulate/articles/python_packages.html#conda-installation}{reticulate article}.
}
