% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart.R
\name{line_chart_spec}
\alias{line_chart_spec}
\title{Title}
\usage{
line_chart_spec(
  gpos,
  chromosome_field = "seqnames",
  genomic_fields = "pos",
  width = 800,
  height = 180,
  y_field = "counts",
  size = 2,
  title = "Basic Marks: line",
  subtitle = "Tutorial Examples"
)
}
\arguments{
\item{chromosome_field}{}

\item{genomic_fields}{}

\item{width}{}

\item{height}{}

\item{y_field}{}

\item{title}{}

\item{subtitle}{}

\item{gr}{}

\item{bin_size}{}
}
\description{
Title
}
\examples{

# generate example GPos object
library(GenomicRanges)
set.seed(42)

# number of rows per chromosome
n <- 100

gpos <- GPos(
  seqnames = Rle(rep(paste0("chr", 1:22), each = n)),
  pos = rep(sample(250e6, n), each = 22)
)

npos <- length(gpos)
mcols(gpos)$counts <- sample(100, npos, replace=TRUE)


# generate gosling spec and render visualization
spec <- line_chart_spec(gpos)
print(spec)

goslingr(spec)

}
