% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_list.R
\name{print_list}
\alias{print_list}
\title{Print a list as source code}
\usage{
print_list(x, indent = 0)
}
\arguments{
\item{x}{A list.}

\item{indent}{Integer. The level of indentation (used for recursive calls). Default is 0.}
}
\value{
No return value, called for side effects (printing).
}
\description{
This function prints a nested list as formatted R source code.
It's particularly useful when you have a nested R list and want to
generate the R code to recreate or modify it.
}
\examples{

# example GRanges object

library(GenomicRanges)

gr <- GRanges(
  seqnames = paste0('chr', c(1:22, 'X', 'Y')),
  peak = rnorm(24, mean = 500, sd = 100),
  ranges = IRanges(
    start = rep(0, 24),
    end = rep(10, 24)
  )
)

# create a bar chart spec
spec <- bar_chart_spec(gr)

# print it for copying/editing in R
print(spec)

}
